GET POST

+++ src/main/kotlin/com/dounine/fastdoc/core/rep/LengthCallback.kt
@@ -2,6 +2,6 @@ package com.dounine.fastdoc.core.rep
 
 interface LengthCallback {
 
-    fun result():Int
+    fun result():Long
 
 }
